Move Page type to address_map module
